# openapi/schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_userLogin",
        "parameters": [
          {
            "name": "x-auth-flow",
            "in": "header",
            "description": "The authentication flow to use. 'web' should be specified for web applications",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Logs in a user and sets the refresh token cookie.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login/admin": {
      "post": {
        "operationId": "AuthController_adminLogin",
        "parameters": [
          {
            "name": "x-auth-flow",
            "in": "header",
            "description": "The authentication flow to use. 'web' should be specified for web applications",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Logs in an admin user and sets the refresh token cookie.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login/otp": {
      "post": {
        "operationId": "AuthController_loginOtp",
        "parameters": [
          {
            "name": "x-auth-flow",
            "in": "header",
            "description": "The authentication flow to use. 'web' should be specified for web applications",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Executes the OTP login stage.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "x-auth-flow",
            "in": "header",
            "description": "The authentication flow to use. 'web' should be specified for web applications",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Refreshes the user's access token and sets a new refresh token cookie.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logs out the user by clearing the refresh token cookie.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getUserProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieves the profile of the currently authenticated user.",
        "tags": [
          "Auth"
        ]
      },
      "put": {
        "operationId": "AuthController_updateUserProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Updates the profile of the currently authenticated user.",
        "tags": [
          "Auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_deleteUserProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletes the profile of the currently authenticated user.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile/picture": {
      "post": {
        "operationId": "AuthController_updateProfilePicture",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationUserDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Updates the profile picture of the currently authenticated user.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signup/customer": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Registers a new customer account.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Requests a password reset for a user.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password/update": {
      "post": {
        "operationId": "AuthController_updatePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordUpdateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Updates the password for a user.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password/change": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changes the password for the currently authenticated user.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/otp": {
      "put": {
        "operationId": "AuthController_switchOtp",
        "parameters": [
          {
            "name": "value",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enables or disables two-factor authentication for the user.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/email": {
      "post": {
        "operationId": "AuthController_requestEmailUpdate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Requests an email update for the currently authenticated user.",
        "tags": [
          "Auth"
        ]
      },
      "put": {
        "operationId": "AuthController_confirmEmail",
        "parameters": [
          {
            "name": "x-auth-flow",
            "in": "header",
            "description": "The authentication flow to use. 'web' should be specified for web applications",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailConfirmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email confirmed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Confirms an email update for the user.",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "EasyMotion API",
    "description": "The EasyMotion API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "Enter your Bearer token"
      }
    },
    "schemas": {
...
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user",
            "example": "admin@easymotion.it"
          },
          "password": {
            "type": "string",
            "description": "Password of the user",
            "example": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "description": "User related data",
            "allOf": [
              {
                "$ref": "#/components/schemas/BaseAuthUserDto"
              }
            ]
          },
          "tokens": {
            "description": "Access token and, eventually, refresh token",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccessTokenDto"
              }
            ]
          },
          "requiresOtp": {
            "type": "boolean",
            "description": "Indicates if secondary OTP step is required"
          }
        },
        "required": [
          "requiresOtp"
        ]
      },
      "OtpLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email used for first login step"
          },
          "otp": {
            "type": "string",
            "description": "One-time password"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token (no need to specify in case of a web auth flow)"
          }
        }
      },
...
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier (UUID)",
            "example": "b3bf4d18-8dd0-43a1-b1da-fd3f7b9553a1"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "nullable": true,
            "description": "User middle name",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "description": "User phone number",
            "example": "+1234567890"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "description": "User birth date",
            "example": "1990-01-01"
          },
          "role": {
            "type": "string",
            "description": "User role in the system",
            "enum": [
              "USER",
              "ADMIN",
              "PHYSIOTHERAPIST"
            ],
            "default": "USER"
          },
          "isEmailVerified": {
            "type": "boolean",
            "description": "Indicates if the user email has been verified",
            "default": false,
            "example": false
          },
          "picturePath": {
            "type": "string",
            "nullable": true,
            "description": "Profile picture path"
          },
          "twoFactorEnabled": {
            "type": "boolean",
            "description": "Indicates if two-factor authentication is enabled",
            "default": false
          },
          "physiotherapist": {
            "nullable": true,
            "description": "Physiotherapist data",
            "allOf": [
              {
                "$ref": "#/components/schemas/PhysiotherapistDto"
              }
            ]
          },
          "patient": {
            "nullable": true,
            "description": "Patient data",
            "allOf": [
              {
                "$ref": "#/components/schemas/PatientDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "isEmailVerified",
          "picturePath",
          "twoFactorEnabled",
          "physiotherapist",
          "patient"
        ]
      },
      "UpdateAuthUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "nullable": true,
            "description": "User middle name",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "description": "User phone number",
            "example": "+1234567890"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "description": "User birth date",
            "example": "1990-01-01"
          },
          "twoFactorEnabled": {
            "type": "boolean",
            "description": "Indicates if two-factor authentication is enabled",
            "default": false
          },
          "physiotherapist": {
            "nullable": true,
            "description": "Physiotherapist data",
            "allOf": [
              {
                "$ref": "#/components/schemas/PhysiotherapistDto"
              }
            ]
          },
          "patient": {
            "nullable": true,
            "description": "Patient data",
            "allOf": [
              {
                "$ref": "#/components/schemas/PatientDto"
              }
            ]
          }
        }
      },
      "ApplicationUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier (UUID)",
            "example": "b3bf4d18-8dd0-43a1-b1da-fd3f7b9553a1"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "nullable": true,
            "description": "User middle name",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "description": "User phone number",
            "example": "+1234567890"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "description": "User birth date",
            "example": "1990-01-01"
          },
          "role": {
            "type": "string",
            "description": "User role in the system",
            "enum": [
              "USER",
              "ADMIN",
              "PHYSIOTHERAPIST"
            ],
            "default": "USER"
          },
          "isEmailVerified": {
            "type": "boolean",
            "description": "Indicates if the user email has been verified",
            "default": false,
            "example": false
          },
          "lastLogin": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user last login",
            "example": "2025-12-31T23:59:59.999Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of when the user was created",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user last update",
            "example": "2025-01-10T10:00:00.000Z"
          },
          "picturePath": {
            "type": "string",
            "nullable": true,
            "description": "Profile picture path"
          },
          "twoFactorEnabled": {
            "type": "boolean",
            "description": "Indicates if two-factor authentication is enabled",
            "default": false
          },
          "physiotherapist": {
            "nullable": true,
            "description": "Physiotherapist data",
            "allOf": [
              {
                "$ref": "#/components/schemas/PhysiotherapistDto"
              }
            ]
          },
          "patient": {
            "nullable": true,
            "description": "Patient data",
            "allOf": [
              {
                "$ref": "#/components/schemas/PatientDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "isEmailVerified",
          "lastLogin",
          "createdAt",
          "updatedAt",
          "picturePath",
          "twoFactorEnabled",
          "physiotherapist",
          "patient"
        ]
      },
      "SignUpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the new user",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password of the user",
            "example": "password"
          },
          "repeatedPassword": {
            "type": "string",
            "description": "Repeated password of the user",
            "example": "password"
          },
          "firstName": {
            "type": "string",
            "description": "First name of the new user",
            "example": "John"
          },
          "middleName": {
            "type": "string",
            "description": "Middle name of the new user (optional)",
            "example": "Alexander"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the new user",
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number of the new user (optional)",
            "example": "+123456789"
          },
          "birthDate": {
            "type": "string",
            "description": "Birth date in ISO format (YYYY-MM-DD) (optional)",
            "example": "1990-01-01"
          }
        },
        "required": [
          "email",
          "password",
          "repeatedPassword",
          "firstName",
          "lastName"
        ]
      },
      "EmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user",
            "example": "user@easymotion.it"
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}